+++ tests/parser/test_parser.py
@@ -4,6 +4,8 @@ import os
 
 import pytest
 
+import fonduer
+from fonduer.parser import Parser
 from fonduer.parser.lingual_parser import SpacyParser
 from fonduer.parser.models import Document
 from fonduer.parser.parser import ParserUDF, SimpleParser
@@ -13,6 +15,15 @@ from fonduer.parser.preprocessors import (
     TextDocPreprocessor,
     TSVDocPreprocessor,
 )
+
+DB = "parser_test"
+if "CI" in os.environ:
+    CONN_STRING = (
+        f"postgresql://{os.environ['PGUSER']}:{os.environ['PGPASSWORD']}"
+        + f"@{os.environ['POSTGRES_HOST']}:{os.environ['POSTGRES_PORT']}/{DB}"
+    )
+else:
+    CONN_STRING = f"postgresql://127.0.0.1:5432/{DB}"
 
 
 def get_parser_udf(
@@ -812,3 +823,110 @@ def test_parser_no_image():
 
     # Check that doc has no figures
     assert len(doc.figures) == 0
+
+
+def test_various_file_path_formats():
+    """Test the parser with various file path formats."""
+    session = fonduer.Meta.init(CONN_STRING).Session()
+
+    def parse(docs_path, pdf_path):
+        # Preprocessor for the Docs
+        doc_preprocessor = HTMLDocPreprocessor(docs_path)
+
+        # Create an Parser and parse the documents
+        corpus_parser = Parser(
+            session,
+            parallelism=1,
+            structural=True,
+            lingual=True,
+            visual=True,
+            pdf_path=pdf_path,
+        )
+
+        corpus_parser.clear()
+        corpus_parser.apply(doc_preprocessor)
+        return corpus_parser
+
+    # Test that doc_path is a file path, and pdf_path is a file path
+    docs_path = "tests/data/html_simple/md.html"
+    pdf_path = "tests/data/pdf_simple/md.pdf"
+
+    corpus_parser = parse(docs_path, pdf_path)
+    docs = corpus_parser.get_documents()
+    assert len(docs) == 1
+    assert docs[0].sentences[0].top is not None
+
+    # Test that doc_path is a file path, and pdf_path is a file path
+    docs_path = "tests/data/html_simple/"
+    pdf_path = "tests/data/pdf_simple/"
+
+    corpus_parser = parse(docs_path, pdf_path)
+    docs = corpus_parser.get_documents()
+    assert len(docs) == 5
+    for doc in docs:
+        # table_span.pdf is not exist, so no coordinate info available
+        if doc.name == "table_span":
+            assert doc.sentences[0].top is None
+        else:
+            assert doc.sentences[0].top is not None
+
+    # Test that doc_path is a directory (no trailing slash), and pdf_path is a
+    # directory
+    docs_path = "tests/data/html_simple"
+    pdf_path = "tests/data/pdf_simple/"
+
+    corpus_parser = parse(docs_path, pdf_path)
+    docs = corpus_parser.get_documents()
+    assert len(docs) == 5
+    for doc in docs:
+        if doc.name == "table_span":
+            assert doc.sentences[0].top is None
+        else:
+            assert doc.sentences[0].top is not None
+
+    # Test that doc_path is a directory, and pdf_path is a directory (no trailing
+    # slash)
+    docs_path = "tests/data/html_simple/"
+    pdf_path = "tests/data/pdf_simple"
+
+    corpus_parser = parse(docs_path, pdf_path)
+    docs = corpus_parser.get_documents()
+    assert len(docs) == 5
+    for doc in docs:
+        if doc.name == "table_span":
+            assert doc.sentences[0].top is None
+        else:
+            assert doc.sentences[0].top is not None
+
+    # Test that doc_path is a directory (no trailing slash), and pdf_path is a
+    # directory (no trailing slash)
+    docs_path = "tests/data/html_simple"
+    pdf_path = "tests/data/pdf_simple"
+
+    corpus_parser = parse(docs_path, pdf_path)
+    docs = corpus_parser.get_documents()
+    assert len(docs) == 5
+    for doc in docs:
+        if doc.name == "table_span":
+            assert doc.sentences[0].top is None
+        else:
+            assert doc.sentences[0].top is not None
+
+    # Test that doc_path is a file path, and pdf_path is a directory
+    docs_path = "tests/data/html_extended/ext_diseases.html"
+    pdf_path = "tests/data/pdf_extended/"
+
+    corpus_parser = parse(docs_path, pdf_path)
+    docs = corpus_parser.get_documents()
+    assert len(docs) == 1
+    assert docs[0].sentences[0].top is not None
+
+    # Test that doc_path is a file path, and pdf_path is a directory (no trailing
+    # slash)
+    docs_path = "tests/data/html_extended/ext_diseases.html"
+    pdf_path = "tests/data/pdf_extended"
+
+    corpus_parser = parse(docs_path, pdf_path)
+    docs = corpus_parser.get_documents()
+    assert len(docs) == 1
+    assert docs[0].sentences[0].top is not None